lease: refactor core lease logic to cleanup locking and retry
The lease function is critical and it was already really long. This
breaks it up into a struct for performing the lease operation, which
also cleans up the locking and makes it possible to retry the whole
process.

The later is needed when we're waiting on a pending container but the
sender closes without sending anything (because they got cancelled)