rpc: Allow resumption of cpabilities
This adds the ability to restore a cpability if the server has
restarted. It requires implementation from the 'root' Interface to
convert previously created InstanceState into an Interface.

In the future, we'll allow the servers to maintain InstanceState in a
server side cache rather than always giving it out to the client to
manage and return back.