Fix activator to use configurable port instead of hard-coded 3000
The activator was hard-coding port 3000 when creating containers, but
some container images like nginx listen on different ports (80). This
change makes use of the existing Config.Port field to properly configure
the container port, falling back to 3000 only if no port is specified.

This fix allows the nginx app example to work properly, as it can now
listen on port 80 as expected.