Move to a testutils implementation of a dev server
This is a hacky copy/paste from commands/dev.go wired up to the
testutils.Registry. The one notable difference is we don't port the
flannel stuff over; we can rely on the existing simple subnet
assignment present in the test registry.

More cleanup to do, but the test is currently passing locally so calling
this a commit checkpoint.