Fix flaky controller tests
The WorkerCount test wasn't providing much value, so I removed it and
enhanced the Lifecycle test to ensure that work is making it through

Running the tests w/ -count 100 caught a panic in the Resync test where
a late Log line was coming through after shutdown. Moving the "shutting
down" logs out of defer seemed to help.