Fix TLS cert validation
Because of the pathway to calling the quic dialing logic, the
tlsConfig.ServerName was not being set (other, more typical pathways
within http3 and websockets set it, but the path we are taking didn't).

As a result, our cert validation was only happening against the IP
address, never against the name. This correctly this and allows for
hostname cert validation.