Overwrite meta-attributes when applying schema
We don't have a way to reconcile things yet and we're still a bit
unstable in terms of schema, so we're going to just jam them in for now.

Using WithOverwrite now only applies the overwrite if the entity is
different than the one stored also, reducing the write load.