Fix bug preventing restarts after project/default changes
The Coordinator would fail to boot after combo.yaml was submitted to a
server, because that file includes a field update for project/default.

We were not catching the Update path on entityserver.Put because the
request did not include an Id, even though db/ident is set in the attrs.
Instead we were relying on an exact-match clause in the response
handling from the Create path.

Changing to entityserver.Client's CreateOrUpdate allows the
Coordinator's project creation to be more flexible

Also added a bunch of logging to make the control flow clearer.

There's more work we can do to clarify the semantics here, but this
fixes the active issue for now.

Fixes MIR-231

Co-authored-by: Teresa Chu <teresa@miren.dev>