Work around bug in yaml parser
Seems that the yaml parser (that is invoked when calling Encode) really
dislikes having a leading newline character! We've openned a bug about
it: https://github.com/yaml/go-yaml/issues/65

This is a fix work around for now, but we're considering moving the
config file body to something opaque rather than a string.