Detect when a sandbox dies on boot, flow error to ingress
We were not terminating the watch of the sandbox status if it
transitioned to other states than running. So when a sandbox died, we
watched it literally until the end of time.

Now we stop as soon as the sandbox transitions to running OR a terminal
state.

Additionally, the httpingress can now detect when this happens and
display a nicery message to the user.