nbd: cleanup large request size handling
So, turns out that nbd servers don't have a way to tell the client to
clamp the maximum payload size. And the linux nbd ioctl/netlink
interfaces don't expose the ability to do that either.

So instead we need to honor the clients size request up to some insane
maximum.