network: Adjust bridge setup
Because of the way linux bridges work, we need to be more careful about
how the bridge is configured early. Otherwise it will enter a status ==
unknown state and not forward traffic.

The trick is the code related to setting the hardware address only after
the bridge has attached interfaces.