sandbox: Only track initial port binding
This code used to use the gvisor syscall monitor to wattch for port
unbinding, so it was effectively free. The version here was FAR from
free, effectively flooding the containers twice a second with new TCP
sessions just to check for continual port binding.

We don't do anything with port unbinding anyway, so this now just waits
for port binding, then stops monitoring.