entity: Minor revision cleanup
Don't log when the revisions are different in the detection case, the
code is working the way it should. But also, be sure to update the
sandbox version when you do an update!

Lastly, because we use the etcd revision as the entity revision, be more
careful about not storing the Revision attr into etcd to reduce
confusion.