controller: fix concurrency entity processing races
The system needs multilpe workers per controller to function currently
(workers occasionally block to wait for a condition). We have always
had a situation where 2 workers could concurrently be processing the
same entity, which is super bad.

This solves the problem by doing a sort of inverted work stealing. If a
worker detects that an event is for an entity that is currently being
processed by another worker, it enqueues that event FOR THAT OTHER
WORKER. Then when a worker finishes it's work, it checks if anyone
dropped additional events for processing and it handles them there.