build: detect identical artifacts
Buildkit can generate an existing image based on the input if it wishes
(it's an optimization!). But when that happened before this, we'd get an
error because the ocireg will refuse to create a new artifact entity
with the same sha256 as an existing one (this is an important property
because we have to lookup artifacts FROM the sha256, so we want there to
be a 1-to-1 between them).

This fixes the issue by running buildkit, but then looking up the
resulting artifact after we build and using that one. This results in
multiple deploys using the same artifact properly.