pool: Add sandbox-labels and sandbox-prefix to sandbox-pool
When a pool is created by the deployer, it's helpful (see later commits
about DNS) to have the sandbox entity tagged with the app it's from.
This lets the deployer setup a set of labels that the pool manager will
dutifully apply to any sandboxes it creates.

Additionally, finally make the entity ID for sandboxes in a pool
meaningfuly, so sandbox_prefix informs the pool manager to name the
sandboxes it creates with a particular naming schema for easily
understanding.