sandbox: patch a sandbox after allocating it's network details
This allows the DNS subsystem to see the address before the sandbox has
fully booted. This is critical because it allows dependent sandboxes to
see the names of their dependent services when they're finally fully
booted.

Without this, service A that needs service B but boots before service B
will get a negative DNS result about service B. With the change, service
A will get the address of service B even when service B hasn't finished
booting.