docker: Let container boot a bit before making UDP requests
The behavior is so bizarre. If we make an http3 request to the forwarded
port before the container has had enough time to startup (note: doesn't
have to actually wait for the server to listen on the port inside the
container, just for the container to actually boot), then docker desktop
removes it's listener for the port!

I have to presume it's some pathalogical condition in docker desktop
itself, but since it's closed source, we can't pin point it.