install: Default to the branch the binary uses (or latest)
The code originally default to main, even when using a proper release
binary. This updates the process to download the base code that matches
the same release by default, while leaving the ability for the user to
specific indicate they want main, latest, or a specific version by
specifying it.